feat: FP32 logits, repetition detection, host KV page fix, context validation (#122)

+++ batchgen/models/minimax/minimax_m25/model.py
@@ -1698,7 +1698,7 @@ class MiniMaxM25(nn.Module):
         )
 
         hidden_states = outputs[0]
-        logits = self.lm_head(hidden_states)
+        logits = self.lm_head(hidden_states).float()
 
         return CausalLMOutputWithPast(
             loss=None,
